+++ camagru/app/auth/signin/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 import styles from '@/styles/Register.module.css'
-import SignInForm from '@/components/SigninForm'
+import SignInForm from '@/components/registration/SigninForm'
 
 export default function SigninForm() {
   return (
