+++ camagru/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
 import './globals.css'
 import Providers from '@/components/Providers'
-import NavbarWrapper from '@/components/NavbarWrapper'
+import NavbarWrapper from '@/components/navbar/NavbarWrapper'
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
