+++ camagru/app/user/[id]/page.tsx
@@ -1,4 +1,4 @@
-import UserProfile from '@/components/UserPage'
+import UserProfile from '@/components/userpage/UserPage'
 
 export default function UserProfilePage() {
   return <UserProfile />
