+++ camagru/styles/Navbar.module.css
@@ -2,7 +2,6 @@
   max-width: 935px;
   margin: 0 auto;
   height: 60px;
-  border-bottom: 1px solid #eee;
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -10,6 +9,12 @@
   position: sticky;
   top: 0;
   z-index: 10;
+}
+
+.divider {
+  width: 100%;
+  height: 1px;
+  background-color: #eee;
 }
 
 .searchForm {
@@ -79,6 +84,18 @@
 
 .navBtn:hover {
   background-color: #f2f2f2;
+  cursor: pointer;
+}
+
+.addBtn {
+  font-size: 1.2rem;
+  border: 1 solid black;
+  border-radius: 20%;
+}
+.addBtn:hover {
+  background-color: transparent;
+  color: #0095f6;
+  border-color: #0095f6;
   cursor: pointer;
 }
