+++ camagru/styles/Profile.module.css
@@ -265,3 +265,23 @@
   display: flex;
   gap: 15px;
 }
+
+.skeletonPost {
+  width: 270px;
+  height: 270px;
+  background-color: #e0e0e0;
+  border-radius: 8px;
+  animation: pulse 1.5s infinite ease-in-out;
+}
+
+@keyframes pulse {
+  0% {
+    background-color: #e0e0e0;
+  }
+  50% {
+    background-color: #f0f0f0;
+  }
+  100% {
+    background-color: #e0e0e0;
+  }
+}
